+++ ErsatzTV.Application/Jellyfin/Mapper.cs
@@ -11,7 +11,7 @@ internal static class Mapper
             jellyfinMediaSource.Connections.HeadOrNone().Match(c => c.Address, string.Empty));
 
     internal static JellyfinLibraryViewModel ProjectToViewModel(JellyfinLibrary library) =>
-        new(library.Id, library.Name, library.MediaKind, library.ShouldSyncItems);
+        new(library.Id, library.Name, library.MediaKind, library.ShouldSyncItems, library.MediaSourceId);
 
     internal static JellyfinPathReplacementViewModel ProjectToViewModel(JellyfinPathReplacement pathReplacement) =>
         new(pathReplacement.Id, pathReplacement.JellyfinPath, pathReplacement.LocalPath);