+++ ErsatzTV.Core/Api/Channels/ChannelResponseModel.cs
@@ -6,6 +6,7 @@ public record ChannelResponseModel(
     int Id,
     string Number,
     string Name,
-    [property: JsonProperty("ffmpegProfile")] string FFmpegProfile,
+    [property: JsonProperty("ffmpegProfile")]
+    string FFmpegProfile,
     string Language,
     string StreamingMode);