+++ ErsatzTV.Core/Emby/EmbyPathInfo.cs
@@ -1,0 +1,8 @@
+﻿namespace ErsatzTV.Core.Emby;
+
+public class EmbyPathInfo
+{
+    public int Id { get; set; }
+    public string Path { get; set; }
+    public string NetworkPath { get; set; }
+}