+++ ErsatzTV.Core/Interfaces/Repositories/IMediaSourceRepository.cs
@@ -26,12 +26,14 @@ public interface IMediaSourceRepository
     Task<List<int>> UpdateLibraries(
         int jellyfinMediaSourceId,
         List<JellyfinLibrary> toAdd,
-        List<JellyfinLibrary> toDelete);
+        List<JellyfinLibrary> toDelete,
+        List<JellyfinLibrary> toUpdate);
 
     Task<List<int>> UpdateLibraries(
         int embyMediaSourceId,
         List<EmbyLibrary> toAdd,
-        List<EmbyLibrary> toDelete);
+        List<EmbyLibrary> toDelete,
+        List<EmbyLibrary> toUpdate);
 
     Task<Unit> UpdatePathReplacements(
         int plexMediaSourceId,