+++ ErsatzTV.Infrastructure/Data/Repositories/JellyfinMovieRepository.cs
@@ -334,12 +334,16 @@ public class JellyfinMovieRepository : IJellyfinMovieRepository
         try
         {
             // blank out etag for initial save in case other updates fail
+            string etag = movie.Etag;
             movie.Etag = string.Empty;
 
             movie.LibraryPathId = library.Paths.Head().Id;
 
             await dbContext.AddAsync(movie);
             await dbContext.SaveChangesAsync();
+
+            // restore etag
+            movie.Etag = etag;
 
             await dbContext.Entry(movie).Reference(m => m.LibraryPath).LoadAsync();
             await dbContext.Entry(movie.LibraryPath).Reference(lp => lp.Library).LoadAsync();