+++ ErsatzTV.Infrastructure/Data/Repositories/PlexTelevisionRepository.cs
@@ -301,12 +301,16 @@ public class PlexTelevisionRepository : IPlexTelevisionRepository
         try
         {
             // blank out etag for initial save in case stats/metadata/etc updates fail
+            string etag = item.Etag;
             item.Etag = string.Empty;
 
             item.LibraryPathId = library.Paths.Head().Id;
 
             await dbContext.PlexShows.AddAsync(item);
             await dbContext.SaveChangesAsync();
+
+            // restore etag
+            item.Etag = etag;
 
             await dbContext.Entry(item).Reference(i => i.LibraryPath).LoadAsync();
             await dbContext.Entry(item.LibraryPath).Reference(lp => lp.Library).LoadAsync();
@@ -326,12 +330,16 @@ public class PlexTelevisionRepository : IPlexTelevisionRepository
         try
         {
             // blank out etag for initial save in case stats/metadata/etc updates fail
+            string etag = item.Etag;
             item.Etag = string.Empty;
 
             item.LibraryPathId = library.Paths.Head().Id;
 
             await dbContext.PlexSeasons.AddAsync(item);
             await dbContext.SaveChangesAsync();
+
+            // restore etag
+            item.Etag = etag;
 
             await dbContext.Entry(item).Reference(i => i.LibraryPath).LoadAsync();
             await dbContext.Entry(item.LibraryPath).Reference(lp => lp.Library).LoadAsync();
@@ -356,6 +364,7 @@ public class PlexTelevisionRepository : IPlexTelevisionRepository
             }
 
             // blank out etag for initial save in case stats/metadata/etc updates fail
+            string etag = item.Etag;
             item.Etag = string.Empty;
 
             item.LibraryPathId = library.Paths.Head().Id;
@@ -371,6 +380,9 @@ public class PlexTelevisionRepository : IPlexTelevisionRepository
 
             await dbContext.PlexEpisodes.AddAsync(item);
             await dbContext.SaveChangesAsync();
+
+            // restore etag
+            item.Etag = etag;
 
             await dbContext.Entry(item).Reference(i => i.LibraryPath).LoadAsync();
             await dbContext.Entry(item.LibraryPath).Reference(lp => lp.Library).LoadAsync();