+++ ErsatzTV.Infrastructure/Emby/Models/EmbyLibraryResponse.cs
@@ -5,4 +5,5 @@ public class EmbyLibraryResponse
     public string Name { get; set; }
     public string CollectionType { get; set; }
     public string ItemId { get; set; }
+    public EmbyLibraryOptionsResponse LibraryOptions { get; set; }
 }