+++ ErsatzTV.Infrastructure/Emby/Models/EmbyPathInfosResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace ErsatzTV.Infrastructure.Emby.Models;
+
+public class EmbyPathInfosResponse
+{
+    public string Path { get; set; }
+    public string NetworkPath { get; set; }
+}